+++ sdk/recoveryservicessiterecovery/arm-recoveryservices-siterecovery/src/models/index.ts
@@ -5057,6 +5057,14 @@ export interface VMwareCbtNicDetails {
   targetIPAddressType?: EthernetAddressType;
   /** Target subnet name. */
   targetSubnetName?: string;
+  /** Source network Id. */
+  testNetworkId?: string;
+  /** Test subnet name. */
+  testSubnetName?: string;
+  /** The test IP address. */
+  testIPAddress?: string;
+  /** The test IP address type. */
+  testIPAddressType?: EthernetAddressType;
   /** Target NIC name. */
   targetNicName?: string;
   /** A value indicating whether this NIC is selected for migration. */
@@ -5077,6 +5085,10 @@ export interface VMwareCbtNicInput {
   isSelectedForMigration?: string;
   /** Target NIC name. */
   targetNicName?: string;
+  /** The test subnet name. */
+  testSubnetName?: string;
+  /** The test static IP address. */
+  testStaticIPAddress?: string;
 }
 
 /** VMwareCbt disk input for update. */
@@ -5085,6 +5097,8 @@ export interface VMwareCbtUpdateDiskInput {
   diskId: string;
   /** The target disk name. */
   targetDiskName?: string;
+  /** A value indicating whether the disk is the OS disk. */
+  isOSDisk?: string;
 }
 
 /** Implements the Alert class. */
@@ -5919,8 +5933,12 @@ export type VMwareCbtEnableMigrationInput = EnableMigrationProviderSpecificInput
   targetResourceGroupId: string;
   /** The target network ARM Id. */
   targetNetworkId: string;
+  /** The selected test network ARM Id. */
+  testNetworkId?: string;
   /** The target subnet name. */
   targetSubnetName?: string;
+  /** The selected test subnet name. */
+  testSubnetName?: string;
   /** The target availability set ARM Id. */
   targetAvailabilitySetId?: string;
   /** The target availability zone. */
@@ -5961,6 +5979,8 @@ export type VMwareCbtUpdateMigrationItemInput = UpdateMigrationItemProviderSpeci
   targetBootDiagnosticsStorageAccountId?: string;
   /** The target network ARM Id. */
   targetNetworkId?: string;
+  /** The test network ARM Id. */
+  testNetworkId?: string;
   /** The list of NIC details. */
   vmNics?: VMwareCbtNicInput[];
   /** The list of disk update properties. */
@@ -6003,6 +6023,8 @@ export type VMwareCbtTestMigrateInput = TestMigrateProviderSpecificInput & {
   recoveryPointId: string;
   /** The test network Id. */
   networkId: string;
+  /** The list of NIC details. */
+  vmNics?: VMwareCbtNicInput[];
 };
 
 /** Single Host fabric provider specific VM settings. */
@@ -7698,15 +7720,15 @@ export type VMwareCbtContainerMappingInput = ReplicationProviderSpecificContaine
   /** Polymorphic discriminator, which specifies the different types this object can be */
   instanceType: "VMwareCbt";
   /** The target key vault ARM Id. */
-  keyVaultId: string;
+  keyVaultId?: string;
   /** The target key vault URL. */
-  keyVaultUri: string;
+  keyVaultUri?: string;
   /** The storage account ARM Id. */
   storageAccountId: string;
   /** The secret name of the storage account. */
-  storageAccountSasSecretName: string;
+  storageAccountSasSecretName?: string;
   /** The secret name of the service bus connection string. */
-  serviceBusConnectionStringSecretName: string;
+  serviceBusConnectionStringSecretName?: string;
   /** The target location. */
   targetLocation: string;
 };