+++ sdk/recoveryservicessiterecovery/arm-recoveryservices-siterecovery/src/models/parameters.ts
@@ -90,7 +90,7 @@ export const $host: OperationURLParameter = {
 export const apiVersion: OperationQueryParameter = {
   parameterPath: "apiVersion",
   mapper: {
-    defaultValue: "2021-11-01",
+    defaultValue: "2022-03-01",
     isConstant: true,
     serializedName: "api-version",
     type: {