+++ sdk/recoveryservicessiterecovery/arm-recoveryservices-siterecovery/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Release History
     
+## 4.1.0 (2022-05-20)
+    
+**Features**
+
+  - Interface VMwareCbtNicDetails has a new optional parameter testIPAddress
+  - Interface VMwareCbtNicDetails has a new optional parameter testIPAddressType
+  - Interface VMwareCbtNicDetails has a new optional parameter testNetworkId
+  - Interface VMwareCbtNicDetails has a new optional parameter testSubnetName
+  - Interface VMwareCbtNicInput has a new optional parameter testStaticIPAddress
+  - Interface VMwareCbtNicInput has a new optional parameter testSubnetName
+  - Interface VMwareCbtUpdateDiskInput has a new optional parameter isOSDisk
+  - Type Alias VMwareCbtEnableMigrationInput has a new parameter testNetworkId
+  - Type Alias VMwareCbtEnableMigrationInput has a new parameter testSubnetName
+  - Type Alias VMwareCbtTestMigrateInput has a new parameter vmNics
+  - Type Alias VMwareCbtUpdateMigrationItemInput has a new parameter testNetworkId
+    
+    
 ## 4.0.0 (2022-01-20)
 
 The package of @azure/arm-recoveryservices-siterecovery is using our next generation design principles since version 4.0.0, which contains breaking changes.